+++ rewatch/src/build.rs
@@ -428,6 +428,12 @@ fn log_config_warnings(build_state: &BuildCommandState) {
                     config::DeprecationWarning::BscFlags => {
                         log_deprecated_config_field(&package.name, "bsc-flags", "compiler-flags");
                     }
+                    config::DeprecationWarning::PackageSpecsEs6 => {
+                        log_deprecated_package_specs_module("es6");
+                    }
+                    config::DeprecationWarning::PackageSpecsEs6Global => {
+                        log_deprecated_package_specs_module("es6-global");
+                    }
                 },
             );
 
@@ -451,6 +457,11 @@ fn log_deprecated_config_field(package_name: &str, field_name: &str, new_field_n
         "The field '{field_name}' found in the package config of '{package_name}' is deprecated and will be removed in a future version.\n\
         Use '{new_field_name}' instead."
     );
+    eprintln!("\n{}", style(warning).yellow());
+}
+
+fn log_deprecated_package_specs_module(module_name: &str) {
+    let warning = format!("deprecated: Option \"{module_name}\" is deprecated. Use \"esmodule\" instead.");
     eprintln!("\n{}", style(warning).yellow());
 }